requirements.txt
yfinance>=0.2.18
pystray>=0.19.4
Pillow>=10.0.0
tzdata>=2024.1
